+++ app/src/main/java/com/example/android_kcs/di/ApiModule.kt
@@ -1,4 +1,4 @@
-package com.example.testcompose2.di
+package com.example.android_kcs.di
 
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -6,5 +6,5 @@ import dagger.hilt.components.SingletonComponent
 
 @Module
 @InstallIn(SingletonComponent::class)
-object NetworkModule {
+object ApiModule {
 }
