+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">testCompose2</string>
+    <string name="app_name">Android-KCS</string>
 </resources>
